Fix wrong value for NEO_COLMASK
ed64ec04 changed NEO_GBR from 0x04 to 0x02 and NEO_COLMASK from 0x05 to 0x08. That value for COLMASK is wrong though.
The purpose of the mask is to only select the bits we want.

The options NEO_RGB, NEO_GRB and NEO_BRG occupy the lower 3 bits.
The correct mask for that is 00000011 -> 0x03.
0x08 would select none of the color bits currently available, defaulting the strip to NEO_RGB mode.