Esp32 bit bang pins above 31 are supported (#829)
must be output pins.
caveat, there seems to be a GCC bug cause the compiler to incorrectly produce code around one of the changes, due to union with a 32bit and 8:24 bit.  The IO registers must be written as 32 bits but the compiler keeps short circuit the code even though it is written to write the whole 32 bits.
The official pinMode() code has a work around in it for this but that doesn't fit into the bitbang code.  This issue maybe already fixed with newer IDF with a newer GCC as this has been reported to work from others.